+++ packages/core/src/domain/configuration.ts
@@ -12,6 +12,7 @@ export const DEFAULT_CONFIGURATION = {
   sampleRate: 100,
   silentMultipleInit: false,
   trackInteractions: false,
+  trackViewsManually: false,
 
   /**
    * arbitrary value, byte precision not needed
@@ -50,8 +51,9 @@ export interface UserConfiguration {
   enableExperimentalFeatures?: string[]
   silentMultipleInit?: boolean
   trackInteractions?: boolean
+  trackViewsManually?: boolean
   proxyHost?: string
-  beforeSend?: (event: any) => void
+  beforeSend?: BeforeSendCallback
 
   service?: string
   env?: string
@@ -66,6 +68,8 @@ export interface UserConfiguration {
   replica?: ReplicaUserConfiguration
 }
 
+export type BeforeSendCallback = (event: any, context?: any) => unknown
+
 interface ReplicaUserConfiguration {
   applicationId?: string
   clientToken: string
@@ -76,7 +80,7 @@ export type Configuration = typeof DEFAULT_CONFIGURATION &
     cookieOptions: CookieOptions
 
     service?: string
-    beforeSend?: (event: any) => unknown
+    beforeSend?: BeforeSendCallback
 
     isEnabled: (feature: string) => boolean
   }
@@ -129,6 +133,10 @@ export function buildConfiguration(userConfiguration: UserConfiguration, buildEn
 
   if ('trackInteractions' in userConfiguration) {
     configuration.trackInteractions = !!userConfiguration.trackInteractions
+  }
+
+  if ('trackViewsManually' in userConfiguration) {
+    configuration.trackViewsManually = !!userConfiguration.trackViewsManually
   }
 
   return configuration