+++ packages/rum-core/src/domain/requestCollection.ts
@@ -34,12 +34,17 @@ export interface RequestCompleteEvent {
   method: string
   url: string
   status: number
-  response?: string
+  responseText?: string
   responseType?: string
   startClocks: ClocksState
   duration: Duration
   spanId?: TraceIdentifier
   traceId?: TraceIdentifier
+  xhr?: XMLHttpRequest
+  response?: Response
+  input?: RequestInfo
+  init?: RequestInit
+  error?: Error
 }
 
 let nextRequestIndex = 1
@@ -69,13 +74,14 @@ export function trackXhr(lifeCycle: LifeCycle, configuration: Configuration, tra
         duration: context.duration,
         method: context.method,
         requestIndex: context.requestIndex,
-        response: context.response,
+        responseText: context.responseText,
         spanId: context.spanId,
         startClocks: context.startClocks,
         status: context.status,
         traceId: context.traceId,
         type: RequestType.XHR,
         url: context.url,
+        xhr: context.xhr,
       })
     }
   })
@@ -101,7 +107,7 @@ export function trackFetch(lifeCycle: LifeCycle, configuration: Configuration, t
         duration: context.duration,
         method: context.method,
         requestIndex: context.requestIndex,
-        response: context.response,
+        responseText: context.responseText,
         responseType: context.responseType,
         spanId: context.spanId,
         startClocks: context.startClocks,
@@ -109,6 +115,9 @@ export function trackFetch(lifeCycle: LifeCycle, configuration: Configuration, t
         traceId: context.traceId,
         type: RequestType.FETCH,
         url: context.url,
+        response: context.response,
+        init: context.init,
+        input: context.input,
       })
     }
   })