+++ packages/rum-core/src/domain/rumEventsCollection/view/viewCollection.ts
@@ -1,24 +1,37 @@
-import { Duration, isEmptyObject, mapValues, ServerDuration, toServerDuration } from '@datadog/browser-core'
+import {
+  Duration,
+  isEmptyObject,
+  mapValues,
+  ServerDuration,
+  toServerDuration,
+  Configuration,
+} from '@datadog/browser-core'
 import { RawRumViewEvent, RumEventType } from '../../../rawRumEvent.types'
-import { LifeCycle, LifeCycleEventType } from '../../lifeCycle'
+import { LifeCycle, LifeCycleEventType, RawRumEventCollectedData } from '../../lifeCycle'
 import { DOMMutationObservable } from '../../../browser/domMutationObservable'
 import { ForegroundContexts } from '../../foregroundContexts'
 import { trackViews, ViewEvent } from './trackViews'
 
 export function startViewCollection(
   lifeCycle: LifeCycle,
+  configuration: Configuration,
   location: Location,
   domMutationObservable: DOMMutationObservable,
-  foregroundContexts: ForegroundContexts
+  foregroundContexts: ForegroundContexts,
+  initialViewName?: string
 ) {
   lifeCycle.subscribe(LifeCycleEventType.VIEW_UPDATED, (view) =>
     lifeCycle.notify(LifeCycleEventType.RAW_RUM_EVENT_COLLECTED, processViewUpdate(view, foregroundContexts))
   )
 
-  return trackViews(location, lifeCycle, domMutationObservable)
+  const shouldTrackViewsAutomatically = !configuration.isEnabled('view-renaming') || !configuration.trackViewsManually
+  return trackViews(location, lifeCycle, domMutationObservable, shouldTrackViewsAutomatically, initialViewName)
 }
 
-function processViewUpdate(view: ViewEvent, foregroundContexts: ForegroundContexts) {
+function processViewUpdate(
+  view: ViewEvent,
+  foregroundContexts: ForegroundContexts
+): RawRumEventCollectedData<RawRumViewEvent> {
   const viewEvent: RawRumViewEvent = {
     _dd: {
       document_version: view.documentVersion,
@@ -67,5 +80,8 @@ function processViewUpdate(view: ViewEvent, foregroundContexts: ForegroundContex
   return {
     rawRumEvent: viewEvent,
     startTime: view.startClocks.relative,
+    domainContext: {
+      location: view.location,
+    },
   }
 }