+++ packages/rum-core/src/index.ts
@@ -9,8 +9,19 @@ export {
   RumResourceEvent,
   RumLongTaskEvent,
 } from './rumEvent.types'
-export { ViewContext, CommonContext } from './rawRumEvent.types'
-export { startRum } from './boot/rum'
+export {
+  ViewContext,
+  CommonContext,
+  RumEventDomainContext,
+  RumViewEventDomainContext,
+  RumErrorEventDomainContext,
+  RumActionEventDomainContext,
+  RumFetchResourceEventDomainContext,
+  RumXhrResourceEventDomainContext,
+  RumOtherResourceEventDomainContext,
+  RumLongTaskEventDomainContext,
+} from './rawRumEvent.types'
+export { startRum } from './boot/startRum'
 export { LifeCycle, LifeCycleEventType } from './domain/lifeCycle'
 export { ParentContexts } from './domain/parentContexts'
 export { RumSession } from './domain/rumSession'