+++ packages/rum-recorder/src/index.ts
@@ -11,6 +11,15 @@ export {
   RumLongTaskEvent,
   RumResourceEvent,
   RumViewEvent,
+  // Events context
+  RumEventDomainContext,
+  RumViewEventDomainContext,
+  RumErrorEventDomainContext,
+  RumActionEventDomainContext,
+  RumFetchResourceEventDomainContext,
+  RumXhrResourceEventDomainContext,
+  RumOtherResourceEventDomainContext,
+  RumLongTaskEventDomainContext,
 } from '@datadog/browser-rum-core'
 
 export { RumRecorderPublicApi as RumGlobal, RumRecorderUserConfiguration } from './boot/rumRecorderPublicApi'